[박준호] 강사 자료실 추가, 파일삭제, 리스트 구현
수정기능 에러 처리중입니다

+++ lms/src/main/java/gd/fintech/lms/vo/LectureArchiveFile.java
@@ -7,7 +7,7 @@ public class LectureArchiveFile {
 	private String lectureArchiveFileUuid;
 	private String lectureArchiveFileOriginal;
 	private int lectureArchiveNo;
-	private int lectureArchiveFileSize;
+	private long lectureArchiveFileSize;
 	private String lectureArchiveFileType;
 	private int lectureArchiveFileCount;
 	private String lectureArchiveFileCreatedate;
